@@ -1,6 +1,6 @@
 import { debounceTime, map } from 'rxjs/operators';
 import { Subject, Observable, from } from 'rxjs';
-import { Injectable } from "@angular/core";
+import { Injectable } from '@angular/core';
 
 /**
  * Allows to access the BIP39 word list.
